updated Git ignore, merged in new header/header.tsx, updated some paths(no capital letters no ü etc.) renamed some css files to scss added QuestionPage.tsx
Note: Only Merge After the new Header Branch was merged
