+++ .gitignore
@@ -22,3 +22,4 @@
 
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
+/node_modules/**
