+++ boredoom-frontend/src/App.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import {MainPage} from "./Pages/MainPage";
+import {MainPage} from "./Pages/MainPage/MainPage";
 import {Route, Switch} from 'react-router'
 import {BrowserRouter} from "react-router-dom";
 import {QuestionPage} from "./Pages/QuestionPage";
@@ -10,7 +10,7 @@ function App() {
         <BrowserRouter>
             <Switch>
                 <Route exact path={"/"} render={() => <MainPage/>}/>
-                <Route exact path={"/fragen"} render={()=> <QuestionPage/>}/>
+                <Route exact path={"/fragenkatalog"} render={()=> <QuestionPage/>}/>
             </Switch>
         </BrowserRouter>
     );
