+++ boredoom-frontend/src/Pages/MainPage/MainPage.scss
@@ -11,12 +11,6 @@
     text-align: center;
     padding-top: 10%;
 }
-.startButton{
-    background: #273661;
-    color: white;
-    padding-left: 50px;
-    padding-right: 50px;
-}
 .StickyFooter{
     border-top: 1px solid #E7E7E7;
     text-align: center;
