+++ boredoom-frontend/src/components/header/header.tsx
@@ -1,11 +1,13 @@
 import "./header.scss";
 import { useCallback } from "react";
 import { Button } from "../button/button";
+import { useHistory } from "react-router-dom";
 
 
 export function Header() {
+    let history = useHistory();
     const getOnClickMenuButton = useCallback((subPath: string) => () => {
-        //history.push(`/${subPath}`);
+        history.push(`/${subPath}`);
     }, []);
 
     return (
@@ -13,8 +15,8 @@ export function Header() {
             <div className="headerLeft">
                 {/*<img className="headerLogo" src={...}/>*/}
                 <div style={{fontSize: "30px", marginRight: "18px", cursor: "pointer"}}>BoreDoom</div>
-                <Button type="header" title="Fragenkatalog" onClick={getOnClickMenuButton("Fragenkatalog")}/>
-                <Button type="header" title="Glückstreffer" onClick={getOnClickMenuButton("Glückstreffer")}/>
+                <Button type="header" title="Fragenkatalog" onClick={getOnClickMenuButton("fragenkatalog")}/>
+                <Button type="header" title="Glückstreffer" onClick={getOnClickMenuButton("glueckstreffer")}/>
                 <Button type="header" title="Merkliste" onClick={getOnClickMenuButton("bookmarks")}/>
                 <Button type="header" title="Kontakt" onClick={getOnClickMenuButton("contact")}/>
             </div>
